+++ go/vax-demo/main.go
@@ -1,7 +1,0 @@
-package main
-
-import "fmt"
-
-func main() {
-    fmt.Println("VAX Protocol Go version is ready!")
-}